+++ src/CDN/Create-CDN-Profile.ps1
@@ -2,7 +2,7 @@
 [CmdletBinding()]
 param (
     [Parameter(Mandatory)][string] $CdnProfileName,
-    [Parameter(Mandatory)][string] $CdnResourceGroup,
+    [Parameter(Mandatory)][string] $CdnResourceGroupName,
     [Parameter()][ValidateSet("Custom_Verizon", "Premium_Verizon", "Standard_Akamai", "Standard_ChinaCdn", "Standard_Microsoft", "Standard_Verizon")][string] $Sku = "Standard_Akamai"
 )
 
@@ -12,6 +12,6 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
-Invoke-Executable az cdn profile create --name $CdnProfileName --resource-group $CdnResourceGroup --sku $Sku
+Invoke-Executable az cdn profile create --name $CdnProfileName --resource-group $CdnResourceGroupName --sku $Sku
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet